ContextManager/requirements.txt
rasa_core==0.9.0a3
rasa_nlu==0.12.2
pymongo==3.6.1
tensorflow==1.4.1
keras==2.0.8
spacy==2.0.11
sklearn_crfsuite==0.3.6